python-converter/requirements_full.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
streamlit==1.51.0
python-pptx==0.6.23
Pillow==10.4.0
pdf2image==1.17.0
